[OpenACC] Overhaul/update READMEs
Reorganize design doc more like the LLVM-HPC paper to make it easier
to find information.  Trim out details of design alternatives that are
better left archived in the LLVM-HPC paper (to be useful to other
developers, this document needs to focus on what we *did* do).  Update
many old details of the design.  Clean up style.