[OpenACC] Make non-const-expr vector_length a warning
Currently, a vector_length clause with a non-constant expression is an
error, but some existing OpenACC programs have that.  After this
patch, it's a warning and the clause is discarded, as OpenACC 2.7
sec. 2.5.10 L805-806 permits:

"The implementation may use a different value than specified based on
limitations imposed by the target architecture."