[OpenACC] Do not discard clause args with side effects
Currently, `num_workers` and `vector_length` are sometimes completely
discarded in the translation to OpenMP.  However, if their arguments
have side effects, their arguments shouldn't be discarded.

In those cases, insert the arguments with casts to `void` before the
OpenMP versions of the directives.  Replace "discard" with "ignore" in
related diagnostics so it doesn't imply the clauses are entirely
thrown away in the translation.