[OpenACC][NFC] Move analysis wrapping simd in parallel for
Currently, we record the partitioning of the parent `acc loop` in each
`acc loop` AST node when we construct it, and translation to OpenMP
then reads that information to determine whether an `omp simd` must be
wrapped in an `omp parallel for num_threads(1)`.  This patch instead
tracks the parent OpenMP loop-related directive during translation to
OpenMP.

There are several reasons to do this.  First, this is really an OpenMP
issue, so it's cleaner to track OpenMP directives not OpenACC
directives.  Second, it's easy to do during translation, where we're
already tracking other inherited attributes (specifically,
`num_workers` and `vector_length`) in a similar manner.  Finally, when
we implement implicit `gang` clauses, the parent partitioning might
change at the end of the construction of an `acc parallel` construct,
but the above analysis is currently before that whereas this patch
moves it afterward.