[OpenACC] Fix num_workers var creation for acc loop auto worker
In the OpenMP translation, don't insert a variable holding the value
of a non-constant-expression `num_workers` for the sake of `acc loop
auto worker`, which Clacc currently translates as `acc loop seq`.
That is, check the result of the loop partitioning computation not the
mere presence of a `worker` clause.