[OpenACC][NFC] Extend and clarify ACCPartitioningKind
Instead of just recording implicit `independent`, explicit
`independent`, or `seq`, extend to record `auto` and to record whether
any of these is implicit, explicit, or computed.  Moreover, be more
rigorous with asserts and documentation to clarify the process of
setting implicit clauses, then explicit clauses if any, and then
computed partitionability if any when have auto or independent.  Don't
forget explicit `gang`, `worker`, and `vector` clauses just because
`auto` has become `seq`.

Recording `auto` and clarifying the process clarifies the code that
analyzes the explicit clauses.  Distinguishing explicit `seq` from
computed `seq` will be necessary in a future patch to fix an issue
with predetermined shared for loop control variables.  Remembering
explicit `gang`, `worker`, and `vector` will be necessary in a future
patch to fix validation of these clauses against the same clauses in
enclosing loops.