[OpenACC] Fix check of partition levels against parent auto
Currently, it fails to complain if a loop construct has `gang`,
`worker`, or `vector` that is incompatible with `gang`, `worker`, or
`vector` on the parent loop construct if the latter has `auto`.  Fix
that by checking for the relevant clauses not the partitioning, which
is removed by conversion of `auto` to `seq`.