[OpenACC] Shared loop control var only for explicit seq
In all cases of sequential execution of an `acc loop`, Clacc currently
makes a loop control variable implicitly shared if it is assigned but
not declared in the init of the attached `for` loop.  That's fine in
the case of an explicit `seq` clause.  However, it's confusing in the
following cases because the OpenACC programmer cannot easily predict
sequential execution (or will not able to easily predict it once Clacc
evolves a descriptive interpretation of OpenACC) and thus cannot
easily determine data sharing for the variable: (1) if `auto` is
present, or (2) if none of `gang`, `worker`, or `vector` are present.

This patch sets implicit shared for such a loop control variable only
if `seq` is explicit.  Otherwise, the variable is predetermined
`private`, as for partitioned loops already.

Besides the benefit to users, this change will also facilitate
implementing implicit `gang` clauses, which, only before this patch,
would influence the data sharing of loop control variables (that's
true for case #2 above even with Clacc's current prescriptive OpenACC
interpretation).  The problem would have been that the data sharing
analysis runs after parsing each construct whereas the implicit `gang`
clause analysis must run after an entire loop nest and would thus
sometimes require fixups to the data sharing analysis.  This patch
removes the dependency between those analyses and thus the need for
fixups.