[OpenACC][NFC] Clean up implicit reduction for acc parallel loop
That is, rather than adding it to both the `acc parallel` and `acc
loop`, add it only to the `acc loop`, and rely on the normal
computation of implicit gang reduction to add it to the `acc
parallel`.  This better follows the logic in the OpenACC spec.  Before
we supported the `copy` clause, that wasn't possible because it's the
`copy` clause implied by a reduction on an `acc parallel loop` that
means some such reductions are for gang-shared variables and thus
imply gang reductions.