[OpenACC] Diagnose reduction conflicts in loop nests
There were already diagnostics for conflicts among explicit and
implicit gang reductions, but there were no diagnostics for conflicts
among reductions from different levels of a loop nest.  The latter is
based on the restriction from OpenACC 2.7 sec. 2.9.11 L1580-1581.

This patch removes the implementation for the old diagnostic from
`ImplicitGangReductionAdder`, and it implements both diagnostics in a
unified manner in the new `NestedReductionChecker`.  Not only is that
good for users, but it also simplifies `ImplicitGangReductionAdder`