[OpenACC][NFC] Gang-partitioned loop's reduction implies copy
Without this patch, a gang-partitioned loop's reduction implies an
`acc parallel` reduction, which implies a `copy` clause, just like an
explicit reduction on the `acc parallel` would.  Adjust this to be the
other way around, as in the design doc: the gang-partitioned loop's
reduction implies the `copy` clause, which makes the reduction
variable gang-shared and thus causes the loop reduction to imply the
`acc parallel` reduction.  Thus, compute implicit base DSAs before
instead of after implicit gang reductions.  The new logic better
reflects current proposals for the OpenACC spec after 2.7.