[OpenACC][NFC] Prepare for r370044 from master
r370044 asserts that statement nodes in the AST are trivially
destructible, but `ACCLoopDirective` is not because it has a
`DenseMap` member for loop control variables.  Replace that with a
`MutableArrayRef<VarDecl *>` referencing a new array stored in the
node's allocation after the clause pointers and the associated
statement pointer.  Change callers to convert to and from `DenseMap`
as needed.