[OpenACC] Diagnose pre private conflicts on combined dirs
Without this patch, Clacc diagnoses conflicts between explicit private
and any other data clause for the same variable.  It also diagnoses a
reduction for a loop control variable, which is predetermined private.
Clacc diagnoses these cases for combined and non-combined directives.

This patch extends the former diagnostic to predetermined private (for
a loop control variable).  The only effect is for conflicts with data
clauses on a combined directive because the only data clauses
supported by a non-combined acc loop are private (doesn't conflict)
and reduction (already diagnosed for loop control variables).