[OpenACC] Model each DA as either a DMA or DSA
This replaces Clacc's old model, which classified each DA as either a
base DA or reduction.  The new model seems to simplify and clarify
documentation, discussion, and implementation.  It's based on OpenMP's
model, suggesting that it's been vetted, will be more recognizable to
potential Clacc contributors, and should facilitate translation from
OpenACC to OpenMP.