[OpenACC] Reduction implies copy even if var is unreferenced
Without this patch, Clacc did not add an implied copy clause for a
reduction var if the var was not referenced within the parallel
construct.  With this patch, it does.  This clarification will appear
after OpenACC 3.0.  Affected cases are expected to be uncommon in real
applications.

However, a more important clarification is that data clauses on
enclosing data constructs also do not suppress a copy clause implied
by a reduction.  This patch prepares for that behavior in the upcoming
data construct implementation.