[OpenACC] Fix var list clause end location
Without this patch, the end location for any OpenACC clause accepting
a variable list was incorrectly recorded as the start of the next
token.  With this patch, it's recorded as the closing ')' location
instead.  This patch matches a similar fix in OpenMP clause parsing
that went in (dbc72c9009a3b) after I copied that implementation.

As far as I can tell, the clause end location isn't currently exposed
in any way, so this change is actually NFC for now.  However, it will
be exposed in diagnostics for the upcoming present clause
implementation.