[Clacc][CI] Fix check-libomptarget fails on ExCL
The problem was that clang was being built so that it directed the
linker to find the system libs rather than the libs for the gcc
selected via `module load gnu/9.1.0`.  Linking often failed when C++
libs were required because of the following issue:

<https://gcc.gnu.org/onlinedocs/libstdc++/manual/using_dual_abi.html>

This affected any compiled program, including `check-libomptarget`
test programs, that links against `libomptarget.so` because that uses
C++ libs.