[OpenACC] Prototype OpenMP TR8 `present` map type modifier
The OpenMP TR8 `present` map type modifier is needed to support the
OpenACC `present` clause, which will be implemented in a following
commit.  Currently, all testing of OpenMP TR8 `present` is via testing
for OpenACC `present`.  That is, OpenMP TR8 `present` is not well
tested outside of translations from OpenACC to OpenMP, so it is not
yet recommended for general use in OpenMP code.