[OpenACC] Prototype OpenMP extension: no_alloc map type modifier
The `no_alloc` map type modifier is an original OpenMP extension for
the sake of supporting OpenACC's `no_create` clause, which will be
implemented in a following commit.  Currently, all testing of
`no_alloc` is via testing for `no_create`.  That is, `no_alloc` is not
well tested outside of translations from OpenACC to OpenMP, so it is
not yet recommended for general use in OpenMP code.