[OpenACC] Improve -fopenacc-print for _Pragma and executable directives
Due to the way Clang tracks locations for the `_Pragma` operator, its
end location cannot be rewritten.  An executable directive (the new
`acc update` is the first) has no associated statement to use for an
end location, so an executable directive in a `_Pragma` operator
directive cannot be rewritten.  Without this patch, the diagnostic
claims the directive appears in a macro in this case.  This patch uses
a better diagnostic for that case.

This patch also extends `-fopenacc-print` testing for executable
directives and updates the status doc.