[OpenACC] Rename -fopenacc-present-omp arguments
Instead of `present` and `alloc`, name them `present` and
`no-present`.  This change is not backward compatible, but we're not
aware of anyone currently using this command-line option.

The new names are more consistent with all the other
`-fopenacc-<ACC-FEATURE>-omp=X|no-X` options Clacc supports now.

Moreover, the new names make more sense given that the `alloc` map
type is always used, and this option and
`-fopenacc-structured-ref-count-omp=hold|no-hold` each control the
inclusion of additional modifiers.