[OpenACC] Rename -fopenacc-no-create-omp arguments
Instead of `no_alloc` and `alloc`, name them `no_alloc` and
`no-no_alloc`.  This change is not backward compatible, but we're not
aware of anyone currently using this command-line option.

The new names are more consistent with all the other
`-fopenacc-<ACC-FEATURE>-omp=X|no-X` options Clacc supports now.

Also, fix misspelling of the `no_alloc` argument as `no-alloc` in the
status doc.