[OpenACC] Prototype OpenMP extension: omp_target_range_is_present
The `omp_target_range_is_present` function is an original OpenMP
extension for the sake of supporting OpenACC's `acc_is_present`
function, which will be implemented in a following commit.  Currently,
all testing of `omp_target_range_is_present` is via testing for
`acc_is_present`.  That is, `omp_target_range_is_present` is not well
tested outside of Clacc's OpenACC runtime, so it is not yet
recommended for general use in OpenMP code.