[OpenACC] Prototype OpenMP extensions: omp_target_map_(to|from|from_delete|alloc|release|delete)
These routines are original OpenMP extension for the sake of
supporting OpenACC's `acc_copyin`, `acc_copyout`,
`acc_copyout_finalize`, `acc_create`, `acc_delete`, and
`acc_delete_finalize` functions, which will be implemented in a
following commit.  For now, all testing of these OpenMP routines will
be via testing of the OpenACC routines.  That is, the OpenMP routines
are not well tested outside of Clacc's OpenACC runtime, so they are
not yet recommended for general use in OpenMP code.

This includes a fix for the behavior of the `hold` modifier when both
`from` and `delete` modifiers are specified, as in the case of
`omp_target_map_from_delete`.  (It doesn't appear possible to combine
those in an `omp target exit data` directive currently.)  Without this
patch, the problem is that `targetDataEnd` computes `DelEntry = true`
due to the `delete` modifier without checking that the hold reference
count is zero, so it erroneously transfers data.  However, it
fortunately doesn't deallocate afterward because it then checks the
total reference count.  With this patch, `DelEntry` is computed
correctly so the erroneous transfer doesn't occur.