[OpenACC][NFC] Simplify FOREACH_CASE idiom in lit tests
That is, make it like the occurrence in the recent
`openmp/libacc2omp/test/prof/api/device-init.c`.  Specifically, use
lit `%for cases` to generate the `FOREACH_CASE` definition so that you
don't have to maintain the list of cases also in the source code, and
so you don't need code to verify that the two lists are in still sync.