[OpenACC] Prototype OpenMP extension: omp_get_mapped_hostptr
`omp_get_mapped_hostptr` is an original OpenMP extension for the sake
of supporting OpenACC's `acc_hostptr`, which will be implemented in a
following commit.  For now, all testing of `omp_get_mapped_hostptr`
will be via testing for `acc_hostptr`.  That is,
`omp_get_mapped_hostptr` is not well tested outside of Clacc's OpenACC
runtime, so it is not yet recommended for general use in OpenMP code.