[OpenACC] Work around max recursion depth for long LIT scripts
That is, avoid the following python error when using LIT's internal
shell for scripts with many RUN lines:

```
RecursionError: maximum recursion depth exceeded
```

I have not identified any upstream test with enough RUN lines to
reproduce this problem, so it seems it's likely only when using
Clacc's `%for` lit extension to generate RUN lines.

In `llvm/utils/lit/lit/TestRunner.py`, `executeScriptInternal`
assembles the RUN lines with a tree of `ShUtil.Seq(cmd, '&&', c)`.
Without this patch, the problem is that the tree is then executed
recursively by `_executeShCmd`.  This patch executes it iteratively
instead.