[OpenACC] Prototype OpenMP extensions: omp_target_update_(to|from)
These routines are original OpenMP extensions for the sake of
supporting OpenACC's `acc_update_device` and `acc_update_self`, which
will be implemented in a following commit.  For now, all testing of
these OpenMP routines will be via testing of the OpenACC routines.
That is, the OpenMP routines are not well tested outside of Clacc's
OpenACC runtime, so they are not yet recommended for general use in
OpenMP code.