[OpenACC] Prototype OpenMP extension: omp_{set,clear}_source_info
These routines are original OpenMP extensions to support reporting
OpenACC Runtime Library routines in OpenACC Profiling Interface
callbacks.  Currently, OpenACC routines are identified for this
purpose based on the OpenMP routines that they call and the info thus
passed via OMPT callbacks.  However, some OpenACC routines remaining
to be implemented (acc_memcpy_*) cannot be identified uniquely in this
manner, and it's possible some of the OpenMP extension routines we're
already calling will be altered during upstreaming/standardization.
Thus, in the future, OpenACC routines will call omp_set_source_info to
override the source info set by OpenMP routines and directives.

For now, all testing of these OpenMP extensions will be via testing of
the OpenACC runtime.  That is, the OpenMP extensions are not well
tested outside of Clacc's OpenACC runtime, so they are not yet
recommended for general use in OpenMP code.