[OpenACC] Fix OpenMP declare variant for arch(x86_64)
Without this patch, Clang accepts the following but fails to generate
variants for x86_64:

```
 #pragma omp begin declare variant match(device = {arch(x86_64)})
```

For the following, Clang correctly produces an error diagnostic
stating that "x86-64" is not valid but "x86_64" is:

```
 #pragma omp begin declare variant match(device = {arch("x86-64")})
```

In the first case, the problem is that, after validating the string
against `OMPKinds.def`, `llvm::Triple::getArchTypeForLLVMName` is
called to convert the string to an `llvm::Triple::ArchType`, but that
recognizes only `x86-64` without this patch.

This patch extends `llvm::Triple::getArchTypeForLLVMName` to also
recognize `x86_64`.  All other callers also validate the string
beforehand, so this patch doesn't appear to add `x86_64` to the
strings acceptable by other callers.

This fix is needed for the upcoming `acc_on_device` implementation in
Clacc.  We'll propose it upstream later.