[OpenACC] Finish fixing OpenMP support for device as host
I tried to do this in ae80003a8ae5 but missed the case of
`__tgt_target_data_end_mapper`.  Thus, without this patch, we still
have runtime errors for `acc exit data` and `acc data` when the target
is set to host but there are non-host devices.  Routines like
`acc_copyout` are not affected because they return early for the case
of shared memory.

That patch's upstream version (10c18c69f2a8) does include this fix.

Tests to cover this case will appear in the next commit.