[OpenACC][NFC] Extract an acc2omp-handlers.h interface
It's for functions that are implemented in libacc2omp and that must be
called by the OpenMP runtime for correct OpenACC handling of certain
OpenMP runtime events.  For now, it specifies only the existing
handler for OpenACC Profiling Interface callback registration.  A
later patch will add a handler for device selection based on OpenACC
environment variables.