[Clacc][OpenMP] Fix second debug name from map clause
This patch fixes a bug from upstream's 97e55cfef5b8 (D89802).  For
example, without it, Clang generates x as the debug map name for both
x and y in the following example:

```
 #pragma omp target map(to: x, y)
 x = y = 1;
```

This fix will be necessary to avoid breaking Clacc tests when a future
patch adjusts Clacc's OpenACC Profiling Interface `var_name` support
to use these OpenMP debug names.

We are attempting to upstream this fix in D101564.