[Clacc][OpenMP] Fix installation of libomptarget plugins
56ca700a7633 accidentally removed installation of libomptarget's
generic-elf-64bit plugins (e.g., x86_64 and ppc64le) into the install
directory's top-level lib subdirectory.  As a result, when attempting
to offload to those architectures, the runtime either quietly disabled
offloading or found a (likely incompatible) version of those plugins
in, for example, a system library directory.  As a result, the runtime
sometimes misbehaved, possibly failing.  This patch restores the
proper installation.  Offloading to nvptx64 was not affected by this
bug.