[Clacc][OpenACC][NFC] *ExecutableDirective -> *DirectiveStmt
In OpenACC, "executable directive" refers to directives like `acc
update` not to directives like `acc parallel`, which are called
"constructs" when combined with their associated statements.  However,
without this patch, the class for an AST node representing an
executable directive or a construct is `ACCExectuableDirective`, which
derives from `ExecutableDirective` as does `OMPExecutableDirective`.

In contrast to directives that are declarative in nature, such as `acc
routine`, the common property of OpenACC executable directives and
constructs is that each is a statement that contains a directive
(compare to an "expression statement" or "statement expression").  For
clarity, this patches renames symbols and files according to the
following scheme:

* `ACCExecutableDirective` -> `ACCDirectiveStmt`
* `ExecutableDirective` -> `DirectiveStmt`
    * `ACCDirectiveStmt` and `OMPExecutableDirective` derive from it.
* `ParseOpenACCDeclarativeOrExecutableDirective` ->
  `ParseOpenACCDirectiveStmt`
    * It will not be used for declarative directives when those are
      implemented in the future.

This patch also improves comments associated comments.