[Clacc][OpenACC] Clean up device finalization callbacks
For OpenACC Profiling Interface support and Clacc's OMPT offload
prototype, without this patch, device finalization callbacks are
handled as follows.  Device numbers are recorded in libomp as devices
are initialized.  Finalization callbacks are dispatched back to back
during `ompt_fini` in libomp.

This patch observes that device initialization is already recorded in
libomptarget, so device finalization callbacks should be handled there
instead of libomp.  Specifically, this patch dispatches device
finalization callbacks in `RTLsTy::UnregisterLib` in libomptarget.

A result is that device finalization callbacks can now actually
measure the time to perform device finalization.  However, as
described in the new design doc note, they are currently grouped by
device type, so the measurement is inflated when there are multiple
devices of the same type.