[Clacc][CI] Fix xfails for release builds
`Clang :: Headers/wasm.c` and a series of AMD GPU tests have typically
failed in our release builds in CI.  However, switching to release
builds by default in our CI config revealed the problem is specific to
release builds on the kolds.  That is, we're not seeing them on
explorer, leconte, or oswald02.  Moreover, unlike the others, `Clang
:: Headers/wasm.c` only fails when asserts are turned off.  Update
xfails accordingly, and fix a few kold builds where you botched the
list during the switch.