[Clacc][OpenACC] Organize Clang OpenACC test suite into subdirectories
And move a couple of tests (enter-exit-data-uninit.c and
subarray-errors.c) into a new "directives" subdirectory of
libacc2omp's test suite.  Unlike other tests here, these tests focus
on the runtime's handling of directives or end-to-end testing and do
not include Clang-specific OpenACC checks, such as printing and
dumping.

This reorganization should make it easier to understand the roles of
tests and to run specific groups of tests.  Moreover, it prepares for
later patches that will do for OpenACC directive testing what
29ced4abd22d did for libacc2omp's test suite: create sub test suites
per offload configuration.  The issue is that only some tests here
benefit from a variety of offloading configurations, and now we can
identify which ones easily by the sub directory.