[Clacc][OpenMP][OpenACC] Fix -fopenmp-target-debug
Clacc defaults to compiling OpenMP applications with the new device
RTL, including when that application is the new device RTL.  Without
this patch, that means it defines new-device-RTL-specific globals,
such as `__omp_rtl_debug_kind`, in both the new device RTL and in any
OpenMP application linked with it.  For example:

```
@__omp_rtl_debug_kind = weak_odr local_unnamed_addr addrspace(1) constant i32 0
```

The new device RTL is compiled with optimizations.  Without this
patch, that means 0 is propagated everywhere for
`__omp_rtl_debug_kind`, making it useless to set it to a different
value when building an application later.

To avoid this problem, this patch adds a new Clang command-line option
`-fno-openmp-target-runtime-globals` and uses it when building the new
device RTL.

That then prevents uses of `Level` in
`openmp/libomptarget/DeviceRTL/src/Debug.cpp` from being optimized
away, and that reveals an amdgcn-link error for `Level` when building
applications.  This patch addresses that too.