[Clacc][OpenACC] Require routine directive before function def
WARNING: This patch implements a new restriction: a function
definition must appear within the scope of at least one explicit and
applying routine directive if there is any within the compilation
unit.  Without this patch, Clacc has a feature such that it implicitly
attaches a routine directive to any earlier definition of the
associated function so that the new restriction isn't necessary.  This
patch removes that feature.  The justification is as follows:

* Clacc's implementation of the feature is buggy, as described below.
* OpenACC 3.2 and earlier already specify the new restriction for the
  routine directive form that takes a function name, but the OpenACC
  technical committee agrees it was also intended for the form that
  takes no function name.  Only the latter form is currently supported
  by Clacc.  Clarifications for the specification are under
  discussion.

The cause of the aforementioned bug is that, in Clang's OpenMP
implementation, accelerator codegen happens at each function
definition before the remainder of the compilation unit is parsed, so
later OpenACC routine directives and their OpenMP declare target
translations are not seen.  Without this patch, a few use cases manage
to work as expected anyway:

* Without this patch, if there's an accelerator use of the function in
  the same compilation unit as the function definition, Clang's OpenMP
  implementation generates the accelerator version of the function
  regardless of any OpenACC routine directives or their OpenMP declare
  target translations.  (This behavior suggests there may be a
  straight-forward way to fix the bug if we should later decide the
  feature is worthwhile after all).
* Without this patch, source-to-source mode attaches the required
  OpenMP declare target directives on the function definitions, so
  subsequent OpenMP compilation then generates the accelerator
  version.  Thus, the bug only affects traditional compilation mode.

Again, this patch eliminates the bug by eliminating the feature.