[Clacc][OpenACC][NFC] Remove SemaOpenACC `#define DirStack`
This approach was copied from SemaOpenMP's `DSAStack` early on to hide
the details of the OpenACC directive stack from the rest of Sema.
That is, without this patch, Sema declares the directive stack's type
as `void*` in Sema, and SemaOpenACC's `DirStack` macro expands to a
static cast of that `void*` to the correct type.  This is confusing
when attempting to declare `DirStack` as a local variable, and it
doesn't cleanly extend to additional OpenACC data.

This patch replaces that approach with an `OpenACCDataTy` struct,
which Sema forward declares, and which SemaOpenACC defines.  For now,
it contains only the directive stack, but future patches will extend
it with more data.