[Clacc][OpenACC] Reject dir in func when routine dir implied later
Without this patch, diagnostics about directives appearing in
functions with routine directives are not reported if the routine
directive was implied after the function definition.  This patch fixes
that.

This patch then adjusts `emitNotesForRoutineDirectiveChain` to handle
the possibility that it must recurse to the function whose definition
is currently being parsed and thus might not have an
`ACCRoutineDeclAttr` attached yet.  That possibility appears to be
impossible to produce before this fix.

That adjustment simplifies the interface of
`emitNotesForRoutineDirectiveChain` and requires `HostFunctionInfoTy`
to depend on `DirStackTy`.