[Clacc][OpenACC] Fix error in func after implying own routine dir
Without this patch, if a function f implies its own routine directive
(for example, it calls g in a compute construct and g was previously
defined to call f), and if f later contains an error related to f's
routine directive, then an assert fails when the diagnostics try to
identify f's routine directive because it was never inserted on the
directive stack.

This patch adjusts the way routine directives are tracked.  First,
only explicit routine directives are now stored on the directive
stack, where they are still used to detect invalid nesting of explicit
directives outside any attached function definition.  Second,
`ActOnFinishedFunctionBodyForOpenACC` now uses the routine directive
on the directive stack to immediately attach an `ACCRoutineDeclAttr`
to the `FunctionDecl`.  Thus, everywhere else, `SemaOpenACC` now uses
`ACCRoutineDeclAttr` and never the directive stack to check if a
function has a routine directive, whether implicit or explicit.  This
approach is simpler in that it eliminates subtle code to push and pop
implicit routine directives on the directive stack.  Moreover, it
fixes the above bug.