[Clacc][OpenACC] Fix assert fail if routine dir after use in error
Sometimes Clang records function uses but discards the AST subtrees
that contain them due to errors in the input code.  Those uses then
cannot be found in the AST later in order to report their locations.
Without this patch, the diagnostic for explicit routine directives
after uses fails an assertion when it doesn't find any uses.  This
patch adds fallbacks for that case.