[Clacc][OpenACC] Improve diag wording for routine dir after use/def
Without this patch, diagnostics for routine directives that appear
after uses or definitions of the function have two issues:

- Each leading error diagnostic seems to describe the routine
  directive but gives the location for the definition or a single use.
- The location of the routine directive is repeated as a note for the
  definition and for every use.

This patch produces a leading error diagnostic with the location of
the routine directive once for the definition and once for all uses
together.  Following each are notes for the locations of the
definition or all uses.