[Clacc][OpenACC] Suppress imp gang clause if use gang func
To do so, this patch records a used function's level of parallelism on
the directive stack in the same manner as a loop construct's level of
parallelism.  That also affects how `num_workers` with a non-constant
expression is translated to OpenMP, but we don't yet make use of that
for function uses.

This patch also updates the documentation's discussion of implicit
gang clauses in the OpenACC spec: they were specified in 3.1.  It
extends the notes there on Clacc's implementation to mention function
calls.