[Clacc][OpenACC] Refactor diag for nested loop par level
Make level-of-parallelism diagnostics for nested loops more concise
and more like level-of-parallelism diagnostics for routine directives.

Simplify the implementation of the former.  When multiple
level-of-parallelism clauses conflict (for example, `acc loop gang
worker` cannot be placed within `acc loop worker vector` for multiple
reasons), a harmless side effect of this change is that different
clauses are selected to appear in the diagnostic than before.

This patch prepares for level-of-parallelism diagnostics for orphaned
loops.