[Clacc][OpenACC] Diagnose non-seq function calls on host
That is, if a function has a routine directive with a gang, worker, or
vector clause, complain if it is called directly from a function that
has no routine directive.  The calling function can execute only
outside compute regions, but the called function requires execution
modes (gang-redundant, etc.) that are impossible outside compute
regions.  In the OpenMP translation, host parallelism might have been
achieved, but OpenACC doesn't specify that behavior.