[Clacc][OpenACC] Implement -Wopenacc-omp-ext
It controls all warning options for OpenMP extensions used in
source-to-source mode: `-Wopenacc-omp-update-present`,
`-Wopenacc-omp-map-ompx-hold`, `-Wopenacc-omp-map-present`,
`-Wopenacc-omp-map-ompx-no-alloc`.  To check that it works, extend
tests for each of the above options to check that it's equivalent.

Also, use it in the Clang driver and at various places in the test
suite to avoid maintaining this list of warning options there.